Animater 필터 에서 ChAnimation.cpp 를 헤더와 같이둠
멤버 변수 프로토콜에 맞게끔 이름 변경 및 초기화안해준 멤버 변수 초기화 했음

ChAnimation.cpp
키입력을 AnimationScene 클래스에서 처리하게 했음
Update 에서 히트박스 충돌 처리를 최상단으로 이동시킴
이동시 충돌하면 제자리로 걷는 코드 작성
ChAnimation.h
불필요한 멤버변수 Player, mpTarget 삭제

SceneManager.cpp
HitBoxScene 클래스 삭제 및 관련내용 제거

Character.cpp
Init, Update, Render 에서 사용하지 않는 코드들 제거
사용하지 않는 함수 제거

Command.h
ChAnimation 클래스에 적용할 수 있게 수정함.

GameObject.h
SetPos함수 수정함 // 포스를 설정후 다시 그려주지 않으면 init후 충돌처리한번됨.

AnimationScene.cpp
ChAnimation 클래스에서 키입력 처리하던 것을 옮겨옴
커맨드를 이용해서 키입력을 받고 1p,2p 키를 나눠놨음.
            1p       2p
left        A     VK_LEFT
right       D     VK_RIGHT
WP          G        N
SP          Y        J
WK          H        M
SK          U        K
사용하는 키는 추후에 협의하여 변경해야 함.

Config.h
전역변수 선언된 것 데이터타입과 이름 변경해야함.
enum class 들 프로토콜에 맞게끔 이름 변경

+++ main.cpp
@@ -57,7 +57,7 @@ int APIENTRY WinMain(_In_ HINSTANCE _hInstance, _In_opt_  HINSTANCE _hPrevInstan
 
 	g_mainGame.Release();
 
-	return message.wParam;
+	return (int)message.wParam;
 }
 
 void SetWindowSize(int startX, int startY, int sizeX, int sizeY)
